feat: add atomic components
