+++ src/components/atoms/ButtonPurple/buttonPurple.module.css
@@ -1,0 +1,8 @@
+.button {
+    background-color: rgb(130, 122, 186);
+    color: white;
+}
+
+.button:hover {
+    box-shadow: 2px 2px rgb(65, 61, 93);
+}
