+++ src/components/atoms/ButtonWhite/ButtonWhite.jsx
@@ -1,0 +1,10 @@
+import styles from './buttonWhite.module.css';
+
+export function ButtonWhite({ children, type, onClick }) {
+  return (
+    // eslint-disable-next-line react/button-has-type
+    <button type={type} className={styles.button} onClick={onClick}>
+      {children}
+    </button>
+  );
+}
