+++ src/components/atoms/ButtonWhite/buttonWhite.module.css
@@ -1,0 +1,8 @@
+.button {
+    background-color: white;
+    color: rgb(74, 67, 121);
+}
+
+.button:hover {
+    box-shadow: 2px 2px rgb(65, 61, 93);
+}
