+++ src/components/organisms/Header/Header.module.css
@@ -1,0 +1,10 @@
+.header {
+    display: flex;
+    justify-content: space-between;
+}
+
+.containerAdmission {
+    height: 100%;
+    display: flex;
+    align-items: center;
+}
