+++ src/components/pages/ContactsPage/contacts.module.css
@@ -1,5 +1,4 @@
 .whiteBackground {
     background-color: white;
-    border-radius: 40px;
     box-shadow: 0 0 20px rgba(0, 0, 0, .2);
 }
