+++ src/index.css
@@ -23,8 +23,8 @@ header, footer {
 }
 
 main {
-  padding-bottom: 100px;
-  padding-top: 50px;
+  padding-bottom: 140px;
+  padding-top: 40px;
   min-height: calc(100vh - 200px);
   display: flex;
   box-sizing: border-box;
@@ -37,18 +37,16 @@ button {
   border-radius: 100px;
   cursor: pointer;
   display: inline-flex;
+  justify-content: center;
   font-family: -apple-system, system-ui, system-ui, "Segoe UI", Roboto, "Helvetica Neue", "Fira Sans", Ubuntu, Oxygen, "Oxygen Sans", Cantarell, "Droid Sans", "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Lucida Grande", Helvetica, Arial, sans-serif;
   font-size: 16px;
   font-weight: 600;
-  justify-content: center;
   line-height: 20px;
-  max-width: 480px;
+  /* max-width: 480px; */
   min-height: 40px;
   overflow: hidden;
   padding: 0px 20px;
   text-align: center;
-  background-color: rgb(218, 213, 254);
-  color: rgb(74, 67, 121);
 }
 
 a {
