+++ src/index.js
@@ -3,7 +3,7 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import './index.css';
 import App from './App';
 import ErrorPage from './components/pages/ErrorPage/ErrorPage';
-import { Main } from './components/pages/MainPage/MainPage';
+import { Main } from './components/pages/MainPage/Main';
 import { Signup } from './components/pages/SignupPage/Signup';
 import { Signin } from './components/pages/SigninPage/Signin';
 import { NewSurveyCreating } from './components/pages/NewSurveyCreatingPage/NewSurveyCreating';
